Change js files to functional jsx components
