+++ app/javascript/components/form/fields/firmware_check_list_field.jsx
